requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
yt-dlp==2024.5.27
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
pydantic-settings==2.2.1
